Moving ccl common location
